pyproject.toml
[tool.poetry]
name = "wikideepseek"
version = "1.0.0"
description = "Pasers fow Wikipedia with API created by FastAPI"
authors = ["Sweet_Sempai <chcolatemilk00@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12"
openai = ">=1.84.0,<2.0.0"
httpx = ">=0.28.1,<0.29.0"
black = ">=25.1.0,<26.0.0"
isort = ">=6.0.1,<7.0.0"
bandit = ">=1.8.3,<2.0.0"
pylint = ">=3.3.7,<4.0.0"
python-dotenv = ">=1.1.0,<2.0.0"
bs4 = ">=0.0.2,<0.0.3"
lxml = ">=5.4.0,<6.0.0"
tenacity = ">=9.1.2,<10.0.0"
sqlalchemy = "^2.0.41"
psycopg2-binary = "^2.9.10"
aerich = "^0.9.0"
python-json-logger = "^3.3.0"
celery = "^5.5.3"
eventlet = "^0.40.0"
dramatiq = {extras = ["redis"], version = "^1.18.0"}

[tool.poetry.dependencies.fastapi]
extras = ["standard"]
version = "^0.115.12"

[tool.poetry.dependencies.redis]
extras = ["async"]
version = "^6.2.0"

[tool.poetry.dependencies.uvicorn]
extras = ["standard"]
version = "^0.34.3"

[tool.poetry.dependencies.tortoise-orm]
extras = ["asyncpg"]
version = "^0.25.1"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.2.0"

[tool.black]
line-length = 88
target-version = ["py312"]
skip-string-normalization = true

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true

[tool.aerich]
tortoise_orm = "app.configs.tortoise_config.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
